requirements.txt
py-etherscan-api==0.8.0
requests==2.26.0
